feat(nick) start implementation of nick command 2

+++ Includes/User/User.hpp
@@ -25,7 +25,7 @@ enum BUFFERSTATUS {
 			std::string receivedBuffer;
 
 			int userSocketFd;
-			long lastPingTimestamp;
+			size_t lastPingTimestamp;
 			size_t uniqueId;
 			UserProperties* properties;
 			Channel *currentChannel;
@@ -37,14 +37,14 @@ enum BUFFERSTATUS {
 			std::string getNickname() const;
 			std::string getIpAddr() const;
 			int getUserSocketFd() const;
-			long getLastPingTimestamp() const;
+			size_t getLastPingTimestamp() const;
 			size_t getUniqueId() const;
 			UserProperties getProperties() const;
 			Channel *getCurrentChannel() const;
 			std::vector<Channel *> getChannelList() const;
 			void addChannelToList(Channel *newChannel);
 			void setNickname(const std::string& nickname);
-			void setLastPingTimestamp(long timestamp);
+			void setLastPingTimestamp(size_t timestamp);
 
 			void addToBuffer(const std::string &incomingBuffer);
 			std::string &getReceivedBuffer();
